requirements.txt
flask
python-docx
flask-socketio
eventlet
python-dotenv